requirements.txt
python-dotenv
gradio
openai
langchain
langchain-community
langchain-experimental
langchain-openai
langgraph
pypdf
faiss-cpu
tiktoken
requests
playwright
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
sentence-transformers 